重构data processor
